requirements.txt
# Leka - LekaOS
# Copyright 2020 APF France handicap
# SPDX-License-Identifier: Apache-2.0

ruamel.yaml==0.18.6
